# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a girlfriend rental service API",
        "title": "Rent A Girlfriend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/bookings": {
            "get": {
                "description": "get all booking of a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookings"
                ],
                "summary": "get all booking of a user",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Booking"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Book a date",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookings"
                ],
                "summary": "Book a new date",
                "parameters": [
                    {
                        "description": "Booking Information",
                        "name": "booking",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.BookingRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Booking"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/bookings/{id}": {
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Cancels a booking if it belongs to the authenticated user. Only the boy who made the booking can cancel it.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookings"
                ],
                "summary": "Cancel an existing booking",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Booking ID to cancel",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cancelled booking details",
                        "schema": {
                            "$ref": "#/definitions/models.Booking"
                        }
                    },
                    "400": {
                        "description": "Invalid booking ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Not your booking",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Booking not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/girlfriends": {
            "get": {
                "description": "Retrieves a list of available girls. Can be filtered by date and age. If no filters are provided, returns all available girls.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "girlfriends"
                ],
                "summary": "Get available girls with optional filtering",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Date to check availability (format: 2024-01-01)",
                        "name": "date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by specific age",
                        "name": "age",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of available girls",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Girl"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid age parameter",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/girlfriends/ratings": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Allows users to submit ratings (1-5 stars) and written reviews for girls they have booked",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "girlfriends"
                ],
                "summary": "Submit a rating and review for a girl",
                "parameters": [
                    {
                        "description": "Rating and review details",
                        "name": "rating",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Rating"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created rating details",
                        "schema": {
                            "$ref": "#/definitions/models.GiveRatingRequest"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload or rating value",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Can't rate without a prior booking",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/girlfriends/{id}": {
            "get": {
                "description": "Retrieves detailed information about a girl's profile using their ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "girlfriends"
                ],
                "summary": "Get a specific girl's profile by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Girl ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Girl profile details",
                        "schema": {
                            "$ref": "#/definitions/models.Girl"
                        }
                    },
                    "404": {
                        "description": "Girl not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Booking": {
            "type": "object",
            "required": [
                "booking_date",
                "boy_user_id",
                "girl_user_id",
                "num_of_days",
                "total_cost"
            ],
            "properties": {
                "booking_date": {
                    "type": "string"
                },
                "boy": {
                    "$ref": "#/definitions/models.Boy"
                },
                "boy_user_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2024-01-01T00:00:00Z"
                },
                "girl": {
                    "$ref": "#/definitions/models.Girl"
                },
                "girl_user_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "num_of_days": {
                    "type": "integer"
                },
                "total_cost": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.BookingRequest": {
            "type": "object",
            "required": [
                "booking_date",
                "girl_id",
                "num_of_days"
            ],
            "properties": {
                "booking_date": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "girl_id": {
                    "type": "integer",
                    "example": 1
                },
                "num_of_days": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
        "models.Boy": {
            "type": "object",
            "required": [
                "age",
                "first_name",
                "last_name",
                "user_id"
            ],
            "properties": {
                "age": {
                    "type": "integer",
                    "example": 25
                },
                "bio": {
                    "type": "string",
                    "example": "I love traveling and meeting new people"
                },
                "deleted_at": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2024-01-01T00:00:00Z"
                },
                "first_name": {
                    "type": "string",
                    "example": "John"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "last_name": {
                    "type": "string",
                    "example": "Doe"
                },
                "profile_picture_url": {
                    "type": "string",
                    "example": "https://example.com/profile.jpg"
                },
                "user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "models.Girl": {
            "type": "object",
            "required": [
                "age",
                "daily_rate",
                "first_name",
                "last_name",
                "user_id"
            ],
            "properties": {
                "age": {
                    "type": "integer",
                    "example": 23
                },
                "bio": {
                    "type": "string",
                    "example": "I enjoy meeting new people"
                },
                "daily_rate": {
                    "type": "integer",
                    "example": 100
                },
                "deleted_at": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2024-01-01T00:00:00Z"
                },
                "first_name": {
                    "type": "string",
                    "example": "Jane"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "last_name": {
                    "type": "string",
                    "example": "Doe"
                },
                "profile_picture_url": {
                    "type": "string",
                    "example": "https://example.com/profile.jpg"
                },
                "user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "models.GiveRatingRequest": {
            "type": "object",
            "required": [
                "girl_id",
                "stars"
            ],
            "properties": {
                "girl_id": {
                    "type": "integer",
                    "example": 1
                },
                "review": {
                    "type": "string",
                    "example": "Great experience!"
                },
                "stars": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
...
        "models.Rating": {
            "type": "object",
            "required": [
                "girl_id",
                "stars"
            ],
            "properties": {
                "deleted_at": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2024-01-01T00:00:00Z"
                },
                "girl_id": {
                    "type": "integer",
                    "example": 1
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "review": {
                    "type": "string",
                    "example": "Great experience!"
                },
                "stars": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}